requirements.txt
speechrecognition==3.8.1
google-cloud-texttospeech==2.11.0
google-generativeai==0.7.2
pydub==0.25.1
pyaudio==0.2.14
python-dotenv==0.21.0



